Fix pre-deploy bugs in expanded metrics features
- Cast ROUND(AVG(peak_ccu)) to ::integer in games.js so pg returns a JS
  number instead of a string, ensuring the falsy guard and arithmetic in
  mom_pct are correct
- Reset rankData/hourlyData to null at the start of their fetch effects
  so navigating between games shows a spinner instead of stale data
- Add valueStyle prop to StatBadge; color mom_pct value green/red

+++ backend/src/routes/games.js
@@ -13,11 +13,11 @@ router.get('/:appid', asyncHandler(async (req, res) => {
   const { rows } = await pool.query(
     `SELECT g.appid, g.name, g.header_image, g.last_fetched_at,
             lc.current_ccu, lc.peak_24h, lc.rank,
-            (SELECT ROUND(AVG(peak_ccu))
+            (SELECT ROUND(AVG(peak_ccu))::integer
              FROM daily_peaks
              WHERE appid = g.appid
                AND peak_date >= DATE_TRUNC('month', CURRENT_DATE)) AS this_month_avg,
-            (SELECT ROUND(AVG(peak_ccu))
+            (SELECT ROUND(AVG(peak_ccu))::integer
              FROM daily_peaks
              WHERE appid = g.appid
                AND peak_date >= DATE_TRUNC('month', CURRENT_DATE) - INTERVAL '1 month'
